requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-cors-headers==3.7.0
django-picklefield==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djongo==1.3.6
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonfield==3.1.0
MarkupSafe==2.1.0
mongoengine==0.24.0
numpy==1.22.2
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.1
PyJWT==1.7.1
pymongo==3.12.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.2.4
uritemplate==4.1.1
urllib3==1.26.8
